Refactor out common range parsing code in storages
We made a similar change in Miniflare 2 recently
